[FEAT] 음주 리포트 API 로직 완성

+++ src/main/java/com/umc/puppymode2/domain/report/converter/DrinkReportConverter.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class DrinkReportConverter {
-    public DrinkReportResponseDTO toDto(Integer goal, Integer drinkCount, Integer drinkDays, Integer achievementRate, Integer scoldedCount) {
+    public DrinkReportResponseDTO toDto(Integer goal, Long drinkCount, Long drinkDays, Integer achievementRate, Integer scoldedCount) {
         return DrinkReportResponseDTO.builder()
                 .goal(goal)
                 .drinkRecordCount(drinkCount)
